+++ hahomematic/platforms/custom/__init__.py
@@ -9,12 +9,10 @@ from typing import Final
 from hahomematic.platforms import device as hmd
 from hahomematic.platforms.custom.definition import entity_definition_exists, get_entity_configs
 from hahomematic.platforms.custom.support import CustomConfig
-from hahomematic.support import loop_safe
 
 _LOGGER: Final = logging.getLogger(__name__)
 
 
-@loop_safe
 def create_custom_entity_and_append_to_device(
     device: hmd.HmDevice,
 ) -> None: