+++ hahomematic/platforms/entity.py
@@ -42,7 +42,7 @@ from hahomematic.platforms.support import (
     convert_value,
     generate_channel_unique_id,
 )
-from hahomematic.support import loop_safe, reduce_args
+from hahomematic.support import loop_check, reduce_args
 
 _LOGGER: Final = logging.getLogger(__name__)
 
@@ -229,7 +229,7 @@ class CallbackEntity(ABC):
         if device_removed_callback in self._device_removed_callbacks:
             self._device_removed_callbacks.remove(device_removed_callback)
 
-    @loop_safe
+    @loop_check
     def fire_entity_updated_callback(self, *args: Any, **kwargs: Any) -> None:
         """Do what is needed when the value of the entity has been updated/refreshed."""
         for callback_handler in self._entity_updated_callbacks:
@@ -238,7 +238,7 @@ class CallbackEntity(ABC):
             except Exception as ex:
                 _LOGGER.warning("FIRE_entity_updated_EVENT failed: %s", reduce_args(args=ex.args))
 
-    @loop_safe
+    @loop_check
     def fire_device_removed_callback(self, *args: Any) -> None:
         """Do what is needed when the entity has been removed."""
         for callback_handler in self._device_removed_callbacks:
@@ -648,7 +648,6 @@ class BaseParameterEntity(Generic[ParameterT, InputParameterT], BaseEntity):
             return multiplier
         return 1
 
-    @loop_safe
     @abstractmethod
     def event(self, value: Any) -> None:
         """Handle event for which this handler has subscribed."""
@@ -672,7 +671,6 @@ class BaseParameterEntity(Generic[ParameterT, InputParameterT], BaseEntity):
             )
         )
 
-    @loop_safe
     def write_value(self, value: Any) -> tuple[ParameterT | None, ParameterT | None]:
         """Update value of the entity."""
         old_value = self._value